Add Average True Range Bands, fix bug in ATR, refactoring

+++ indicator/src/main/java/pro/crypto/model/request/BBRequest.java
@@ -20,7 +20,7 @@ public class BBRequest {
 
     private PriceType priceType;
 
-    private int standardDeviationCoefficient;
+    private double standardDeviationCoefficient;
 
     private IndicatorType movingAverageType;
